doc_handler/pyproject.toml
[project]
name = "doc-handler"
version = "0.1.0"
description = "Editor Markdown con IA"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "rich>=13.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["doc_handler"]
addopts = "-v --cov=doc_handler --cov-report=term-missing"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
